Replace the sigmoid with a gaussian in FANN, because FANN is unstable when using a sigmoid
For unexplained reasons, FANN networks sometimes degenerate (and basically learn
the opposite of what they have to learn) when using a sigmoid activation function.

A possible reason is that the output of the networks is sometimes "large" (way above
1 or below -1), and the derivative of the sigmoid is nearly horizontal at those
regions. The gaussian activation does not present this problem and is able to learn
correctly and reliably.