Only train the model with the last N observations of an episode
This comes from the intuition that, when an episode becomes long enough,
the first states in it correspond to "errors" made by the agent and should
not be used for learning (if the agent is busy spinning somewhere or going
towards a wall). Only the few last steps, when the agent either reaches the
goal or fail because it is not at the goal, provide useful information.

This changes dramatically improve the stability of function-approximation-based
Q-Learning and advantage learning