Store the cumulative reward in Episode
This way, even if the episode gets truncated to its maximum size, the total
non-truncated cumulative reward can be recovered and plotted. This allows to
produce plots a bit more informative that plots that a clamped to -100 for 800
time steps and then suddenly achieve much better results. Now we can see the
progressive learning that occurs.